Rename trials to trial so that it doesn't introduce new namespace
