+++ optuna/trial/__init__.py
@@ -1,0 +1,5 @@
+from optuna.trial.base import BaseTrial  # NOQA
+from optuna.trial.fixed import FixedTrial  # NOQA
+from optuna.trial.frozen import FrozenTrial  # NOQA
+from optuna.trial.state import TrialState  # NOQA
+from optuna.trial.trial import Trial  # NOQA
