+++ optuna/trial/fixed.py
@@ -1,8 +1,8 @@
 import datetime
 
 from optuna import distributions
-from optuna.trials.base import BaseTrial
-from optuna.trials.util import _adjust_discrete_uniform_high
+from optuna.trial.base import BaseTrial
+from optuna.trial.util import _adjust_discrete_uniform_high
 from optuna import type_checking
 
 if type_checking.TYPE_CHECKING:
