+++ optuna/trial/util.py
